Checking of products if they are already in Cart. + minor cosmetic defects.
