+++ src/js/addProduct.js
@@ -1,13 +1,14 @@
-import iconsPath from '../icons/icons.svg';
 import refs from './refs.js';
 import Storage from './storage.js';
+import { updateCartOnHeader } from './header.js';
+import { disableBuyProductById } from './checkProducts.js';
 
 const STORAGE_KEY = 'localKey';
 
 refs.allProductsWrapperEl.addEventListener('click', onAddBtnClick);
 
 function onAddBtnClick(e) {
-  if (!e.target.closest('button')) {
+  if (!e.target.closest('button.js-buy-btn')) {
     return;
   }
 
@@ -19,7 +20,7 @@ function onAddBtnClick(e) {
   }
 
   disableBuyProductById(id);
-  increaseCartCounter();
+  updateCartOnHeader();
 }
 
 function saveIdToStorage(id) {
@@ -34,19 +35,4 @@ function saveIdToStorage(id) {
     Storage.save(STORAGE_KEY, selectedProdIds);
     return true;
   }
-}
-
-function disableBuyProductById(id) {
-  const cardsEls = refs.allProductsWrapperEl.querySelectorAll(`[data-product-id='${id}']`);
-  cardsEls.forEach(cardEl => {
-    const button = cardEl.querySelector('button');
-    const useTag = cardEl.querySelector('button use');
-    useTag.setAttribute('href', `${iconsPath}#icon-check`);
-    button.disabled = true;
-  });
-}
-
-function increaseCartCounter() {
-  let cartValue = Number(refs.cartCountSpan.textContent);
-  refs.cartCountSpan.textContent = (++cartValue).toString();
 }
