+++ src/js/onLoad.js
@@ -4,6 +4,10 @@ import { renderPopularProducts } from './popularProduct';
 import { renderProductsCards } from './productsList';
 import { renderDiscountCards } from './discountProducts';
 import { updateCartOnHeader } from './header';
+import { checkProducts } from './checkProducts.js'
+import Storage from './storage.js'
+
+const STORAGE_KEY = 'localKey';
 
 const foodBoutiqueApi = new FoodBoutiqueAPI();
 
@@ -12,8 +16,8 @@ window.addEventListener(`DOMContentLoaded`, onDOMContentLoaded);
 async function onDOMContentLoaded() {
   try {
     console.log('here');
-    updateCartOnHeader()
-    
+    updateCartOnHeader();
+
     let allProducts = await foodBoutiqueApi.fetchProductsByQuery();
     renderProductsCards(allProducts.results, refs.productsListEl);
 
@@ -22,6 +26,8 @@ async function onDOMContentLoaded() {
 
     let discountProducts = await foodBoutiqueApi.fetchDiscount();
     renderDiscountCards(discountProducts, refs.discountProductsEl);
+
+    checkProducts(Storage.load(STORAGE_KEY))
   } catch (err) {
     console.log('err');
   }
